refactor: uriComponents.tsを作成
