+++ src/consts/uriComponents.ts
@@ -7,10 +7,9 @@ export const pagePaths = {
   routeEditorPath: (routeId: string) => {
     return `/routes/${routeId}`;
   },
+  START_MARKER_ICON: "/icons/start_pin.svg",
+  MARKER_ICON: "/icons/fmd_good_black_24dp.svg",
+  GOAL_MARKER_ICON: "/icons/goal_pin.svg",
 };
 
-export const staticFileUrls = {
-  START_MARKER_ICON: "http://localhost:3000/icons/start_pin.svg",
-  MARKER_ICON: "http://localhost:3000/icons/fmd_good_black_24dp.svg",
-  GOAL_MARKER_ICON: "http://localhost:3000/icons/goal_pin.svg",
-};
+export const ORIGIN = "http://localhost:3000";
